исправил опечатку, перенёс функции в utils
