+++ src/mocks/offers.js
@@ -1,3 +1,5 @@
+import {getRandomBoolValue, getRandomNumber} from "../utils";
+
 const cardTitles = [
   `Beautiful & luxurious apartment at great location`,
   `Wood and stone place`,
@@ -19,24 +21,13 @@ const cardPrices = [
   180
 ];
 
-const getRundomBoolValue = () => {
-  const value = Math.random();
-  return value > 0.5;
-};
-
-export const getRandomNumber = (min, max) => {
-  let rand = min + Math.random() * (max + 1 - min);
-  rand = Math.floor(rand);
-  return rand;
-};
-
 function Offer(index) {
   this.name = cardTitles[index];
   this.image = cardImages[index];
   this.price = cardPrices[index];
   this.rating = getRandomNumber(0, 10);
-  this.isPremium = getRundomBoolValue();
-  this.isFavorite = getRundomBoolValue();
+  this.isPremium = getRandomBoolValue();
+  this.isFavorite = getRandomBoolValue();
 }
 
 export const offers = cardTitles.map((it, index) => {
